1981.2: Merge verification indicator component files
# Conflicts:
#	administration/src/bp-modules/applications/__tests__/PreVerifiedQuickIndicator.test.tsx
