+++ administration/src/bp-modules/applications/__tests__/VerificationsIndicator.test.tsx
@@ -1,7 +1,8 @@
 import React from 'react'
 
 import { renderWithTranslation } from '../../../testing/render'
-import VerificationQuickIndicator from '../VerificationsQuickIndicator'
+import { PreVerifiedEntitlementType, preVerifiedEntitlements } from '../PreVerifiedEntitlementType'
+import { PreVerifiedQuickIndicator, VerificationQuickIndicator } from '../VerificationsIndicator'
 import {
   verificationsAwaiting,
   verificationsMixed,
@@ -42,5 +43,18 @@ describe('VerificationQuickIndicator', () => {
     expect(getByTestId(`indicator-${VerificationStatus.Verified}`).textContent).toBe('icon: 2')
     expect(getByTestId(`indicator-${VerificationStatus.Pending}`).textContent).toBe('icon: 0')
     expect(getByTestId(`indicator-${VerificationStatus.Rejected}`).textContent).toBe('icon: 0')
+  })
+})
+
+describe('PreVerifiedQuickIndicator', () => {
+  const cases: [string, PreVerifiedEntitlementType][] = [
+    ['Juleica', preVerifiedEntitlements.Juleica],
+    ['Verein360', preVerifiedEntitlements.Verein360],
+    ['Ehrenzeichen', preVerifiedEntitlements.HonoredByMinisterPresident],
+  ]
+
+  it.each(cases)('should show %s label when type is %s', (expectedLabel, type) => {
+    const { getByText } = renderWithTranslation(<PreVerifiedQuickIndicator type={type} />)
+    expect(getByText(expectedLabel)).toBeTruthy()
   })
 })
